+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/BrowserToolbar.kt
@@ -165,6 +165,10 @@ class BrowserToolbar @JvmOverloads constructor(
         edit.setAutocompleteListener(filter)
     }
 
+    override fun refreshAutocomplete() {
+        edit.refreshAutocompleteSuggestion()
+    }
+
     init {
         addView(display.rootView)
         addView(edit.rootView)